pyproject.toml
[tool.poetry]
name = "forever-rss"
version = "0.1.0"
description = "Your Internet Inbox"
authors = ["Rob van der Leek <5324924+robvanderleek@users.noreply.github.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11,<3.14"
fastapi = {extras = ["standard"], version = "^0.115.8"}
pydantic-settings = "^2.8.0"
requests = "^2.32.3"
pyjwt = "^2.10.1"
sqlalchemy = "^2.0.38"
alembic = "^1.15.1"
psycopg2-binary = "^2.9.10"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
